requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic[email]==2.4.2
pydantic-settings==2.0.3
supabase==2.0.3
python-jose==3.3.0
passlib[bcrypt]
python-dotenv==1.0.0
pytest>=6.2.0
httpx>=0.24.0
jinja2>=3.0.0
pandas>=1.3.0
scikit-learn>=0.24.0
psycopg2-binary
bcrypt==3.2.2
python-multipart==0.0.6
